# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EduHub Backend API
  version: 1.0.0
  description: |
    EduHub Backend is a CRM/SaaS backend for educational centers.
    It includes authentication, users, profiles, roles, permissions,
    students, parents, teachers, groups, lessons, attendance, homework,
    schedules, payments, payroll, reports, files, notifications, imports,
    dashboard and AI chat.
servers:
  - url: http://localhost:8080/api/v1
    description: Local API
security:
  - BearerAuth: []
...
paths:
...
  /branches:
    get:
      tags: [Branches]
      summary: List branches
      responses:
        "200":
          description: Branches returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Branches]
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchRequest'
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /subjects:
    get:
      tags: [Subjects]
      summary: List subjects
      responses:
        "200":
          description: Subjects returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Subjects]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubjectRequest'
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /users:
    get:
      tags: [Users]
      summary: List users
      responses:
        "200":
          description: Users returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      responses:
        "200":
          description: Roles returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /teachers:
    get:
      tags: [Teachers]
      summary: List teachers
      responses:
        "200":
          description: Teachers returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Teachers]
      summary: Create teacher profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /students:
    get:
      tags: [Students]
      summary: List students
      responses:
        "200":
          description: Students returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /parents:
    get:
      tags: [Parents]
      summary: List parents
      responses:
        "200":
          description: Parents returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Parents]
      summary: Create parent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParentRequest'
      responses:
        "201":
          description: Parent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /groups:
    get:
      tags: [Groups]
      summary: List groups
      responses:
        "200":
          description: Groups returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /lessons:
    get:
      tags: [Lessons]
      summary: List lessons
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: group_id
          in: query
          schema:
            type: string
            format: uuid
        - name: teacher_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lessons returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Lessons]
      summary: Create lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /homeworks:
    get:
      tags: [Homework]
      summary: List homeworks
      responses:
        "200":
          description: Homeworks returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Homework]
      summary: Create homework
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Homework created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /schedules:
    get:
      tags: [Schedules]
      summary: List schedules
      responses:
        "200":
          description: Schedules returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Schedules]
      summary: Create schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /payments:
    get:
      tags: [Payments]
      summary: List payments
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: group_id
          in: query
          schema:
            type: string
            format: uuid
        - name: student_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Payments]
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /reports/payments:
    get:
      tags: [Reports]
      summary: Payments report
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - $ref: '#/components/parameters/Format'
        - $ref: '#/components/parameters/Lang'
      responses:
        "200":
          description: Payments report returned or exported
...
  /reports/payroll:
    get:
      tags: [Reports]
      summary: Payroll report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            example: "2026-07"
        - $ref: '#/components/parameters/Format'
        - $ref: '#/components/parameters/Lang'
      responses:
        "200":
          description: Payroll report returned or exported
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      responses:
        "200":
          description: Notifications returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Notifications]
      summary: Create notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /audit-logs:
    get:
      tags: [Audit Logs]
      summary: List audit logs
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: entity_type
          in: query
          schema:
            type: string
        - name: entity_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - name: limit
          in: query
          schema:
            type: integer
            example: 100
        - name: offset
          in: query
          schema:
            type: integer
            example: 0
        - $ref: '#/components/parameters/Format'
        - $ref: '#/components/parameters/Lang'
      responses:
        "200":
          description: Audit logs returned or exported
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    FromDate:
      name: from_date
      in: query
      schema:
        type: string
        format: date
        example: "2026-07-01"
    ToDate:
      name: to_date
      in: query
      schema:
        type: string
        format: date
        example: "2026-07-31"
    Format:
      name: format
      in: query
      schema:
        type: string
        enum: [json, xlsx, pdf, docx]
    Lang:
      name: lang
      in: query
      schema:
        type: string
        enum: [en, ru, kk]
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
...
    CreateBranchRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Skillset
        address:
          type: string
          example: Astana
        phone:
          type: string
          example: "+77000000000"
    CreateSubjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Programming
        description:
          type: string
          example: Programming course
    CreateParentRequest:
      type: object
      required: [full_name]
      properties:
        full_name:
          type: string
          example: Parent Name
        phone:
          type: string
          example: "+77000000001"
        email:
          type: string
          format: email
          example: parent@example.com